+++ Changelog.md
@@ -23,6 +23,7 @@ v2.42.0 (not published yet)
 - UI: Make number of audio tracks on main window optional
 - UI: Re-implement audio track detection and reordering
 - Crop: Make AutoCrop process skippable ([#1411](/../../issues/1411))
+- Qaac: Add "--no-delay" parameter ([#1410](/../../issues/1410))
 - Update tools
     - ...
 - Update AviSynth+ plugins